requirements.txt
discord.py==2.5.2
python-dotenv
playwright
rapidfuzz
beautifulsoup4
requests